remove hotwired (for now)
Turbolinks is just way more trouble than it's worth, at least for now -- I'd rather solve this problem through appropriate caching / JS code splitting.  If we're loading so much shit in the head that it's a problem, honestly we have bigger problems to deal with than anything turbolinks can fix.